src/pdf-splitter/requirements.txt
Flask==2.0.1
google-cloud-storage==1.41.1
gunicorn==20.1.0
pdf2image==1.16.0